It now uses isochronous like it should and gets packets every 10ms for the
control loops nicely. The gyro board code has also been reorganized so it is
much clearer and easier to maintain.

+++ aos/common/control_loop/ControlLoop.cc
@@ -4,8 +4,8 @@ namespace aos {
 namespace control_loops {
 
 time::Time NextLoopTime(time::Time start) {
-  return (start / kLoopFrequency.ToNSec()) *
-      kLoopFrequency.ToNSec() +
+  return (start / static_cast<int32_t>(kLoopFrequency.ToNSec())) *
+      static_cast<int32_t>(kLoopFrequency.ToNSec()) +
       kLoopFrequency;
 }
